+++ src/index.ts
@@ -1,6 +1,6 @@
-import {entity, id, repository, excludeFromIndex, createdAt, updatedAt} from "./decorators";
+import {createdAt, entity, excludeFromIndex, id, repository, updatedAt} from "./decorators";
 import {interfaces} from "./interfaces";
-import {BaseRepository} from "./base_repository";
+import {BaseRepository, PagedBaseRepository} from "./base_repository";
 import {EntityValidationError, Filter, Namespaced, Operator, Order, QueryRequest} from "./types";
 import {TYPES} from "./constants";
 
@@ -13,6 +13,7 @@ export {
   updatedAt,
   interfaces,
   BaseRepository,
+  PagedBaseRepository,
   EntityValidationError,
   Filter,
   QueryRequest,